tx: remove basetx [WIP] [no ci]
